Added behaviour type to active areas

+++ Assets/uAdventure/__Scripts/Core/Loader/Subparsers/SceneSubParser.cs
@@ -98,7 +98,7 @@ namespace uAdventure.Core
 				int destinyX 		= ExParsers.ParseDefault (el.GetAttribute ("destinyX"), int.MinValue), 
 					destinyY 		= ExParsers.ParseDefault (el.GetAttribute ("destinyY"), int.MinValue);
 
-                float destinyScale  = ExParsers.ParseDefault (el.GetAttribute ("destinyScale"), int.MinValue);
+                float destinyScale  = ExParsers.ParseDefault (el.GetAttribute ("destinyScale"), float.MinValue);
 
 				int transitionType 	= ExParsers.ParseDefault(el.GetAttribute("transitionType"), 0),
 					transitionTime 	= ExParsers.ParseDefault(el.GetAttribute("transitionTime"), 0);
